+++ src/chit/mod.rs
@@ -1,6 +1,7 @@
 pub mod details;
 pub mod owners;
 pub mod versions;
+pub mod repo;
 
 mod sources;
 mod extract;